+++ pytrackio/__init__.py
@@ -1,5 +1,5 @@
 from ._registry import _REGISTRY, MetricSummary, MetricsRegistry
-from .core import track
+from .core import track, export_histogram, set_buckets
 from ._timer import timer
 from ._report import report, export_json, export_csv, export_dict
 from .exporters import export_jsonlines
@@ -14,4 +14,4 @@ def get_registry():
 __version__ = "0.10.0"
 __author__ = "Deepanshu"
 __license__ = "MIT"
-__all__ = ["track","timer","counter","report","export_json","export_csv","export_dict","export_jsonlines","get_registry","MetricSummary","MetricsRegistry","__version__"]
+__all__ = ["track","timer","counter","report","export_json","export_csv","export_dict","export_jsonlines","get_registry","MetricSummary","MetricsRegistry","__version__","track", "export_histogram", "set_buckets"]